requirements.txt
asgiref==3.7.2
Django==4.2.3
django-extensions==3.2.3
django-filter==23.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
